+++ src/lib.rs
@@ -2,9 +2,8 @@ pub mod server_config {
     pub mod config;
     pub use config::*;
 
-    use crate::server::Port;
     use crate::server_config::route::Route;
-    use crate::type_aliases::Path;
+    use crate::type_aliases::{Path, Port};
     use std::collections::HashMap;
 
     #[derive(Clone, Debug)]
@@ -18,6 +17,7 @@ pub mod server_config {
 
     pub mod builder {
         use super::*;
+        use crate::type_aliases::Port;
         #[derive(Debug)]
         pub struct ConfigBuilder<'a> {
             pub host: Option<&'a str>,
@@ -84,7 +84,7 @@ pub mod server_config {
     }
 
     pub mod route {
-        use crate::server_config::route::cgi::Cgi;
+        use crate::server::Cgi;
         use crate::type_aliases::{FileExtension, Path};
         use std::collections::HashMap;
 
@@ -111,16 +111,6 @@ pub mod server_config {
                 }
             }
         }
-
-        pub mod cgi {
-            #[derive(Clone, Debug)]
-            pub enum Cgi {
-                Python,
-                PHP,
-                JavaScript,
-                Cpp,
-            }
-        }
     }
 }
 pub mod type_aliases {
@@ -130,29 +120,27 @@ pub mod type_aliases {
     pub type FileExtension<'a> = &'a str;
 }
 
-pub mod client {
+pub mod server {
     pub mod handle;
+
     pub use handle::*;
+    use mio::net::TcpListener;
+    use std::sync::Arc;
 
     pub mod requests;
     pub use requests::*;
 
     pub mod responses;
     pub use responses::*;
-}
-pub mod server {
-    use crate::client::handle_client;
-    use crate::server_config::config::server_config;
+
+    pub mod cgi;
+    pub use cgi::*;
+
+    pub mod routes;
+    pub use routes::*;
+    pub mod start;
     use crate::server_config::ServerConfig;
-    pub use crate::type_aliases::Port;
-
-    use mio::net::TcpListener;
-    use mio::{Events, Interest, Poll, Token};
-    use std::collections::HashMap;
-    use std::io::ErrorKind;
-
-    use std::net::SocketAddr;
-    use std::sync::Arc;
+    pub use start::*;
 
     #[derive(Debug)]
     pub struct Server<'a> {
@@ -166,122 +154,9 @@ pub mod server {
         }
     }
 
-    pub fn servers() -> Vec<Server<'static>> {
-        let mut servers = Vec::new();
-
-        for config in server_config() {
-            let mut listeners = Vec::new();
-            for port in &config.ports {
-                // Create a listener for each port
-                let address = format!("{}:{}", config.host, port);
-                match TcpListener::bind(address.parse::<SocketAddr>().unwrap()) {
-                    Ok(listener) => {
-                        listeners.push(listener);
-                        println!("Server listening on {}", address);
-                    }
-                    Err(e) => eprintln!("Error: {}. Unable to listen to: {}", e, address),
-                }
-            }
-            // Make a server and push it to the servers vector
-            servers.push(Server::new(listeners, config))
-        }
-        servers
-    }
-
+    #[derive(Debug)]
     struct ListenerConfig<'a> {
         listener_index: usize,
         config: Arc<ServerConfig<'a>>,
-    }
-
-    pub fn start(servers: Vec<Server<'static>>) {
-        let mut poll = Poll::new().expect("Failed to create Poll instance");
-        let mut events = Events::with_capacity(128);
-        let mut connections = HashMap::new();
-        let mut token_id = 2; // Start token counting from 2
-        let mut all_listeners = Vec::new(); // Store all listeners
-        let mut listener_configs = Vec::new(); // Store ListenerConfig instances
-
-        for server in servers {
-            let config = Arc::new(server.config);
-
-            for listener in server.listeners {
-                let listener_index = all_listeners.len();
-                all_listeners.push(listener);
-                let token = Token(token_id);
-                token_id += 1;
-
-                poll.registry()
-                    .register(
-                        &mut all_listeners[listener_index],
-                        token,
-                        Interest::READABLE,
-                    )
-                    .expect("Failed to register listener");
-
-                listener_configs.push(ListenerConfig {
-                    listener_index,
-                    config: Arc::clone(&config),
-                });
-            }
-        }
-
-        // Event loop
-        loop {
-            poll.poll(&mut events, None).expect("Poll failed");
-            let mut to_close = Vec::new();
-
-            for event in events.iter() {
-                let token = event.token();
-
-                if let Some(listener_config) = listener_configs
-                    .iter()
-                    .find(|lc| token == Token(lc.listener_index + 2))
-                {
-                    let listener = &mut all_listeners[listener_config.listener_index];
-
-                    // Accept new connections in a loop
-                    loop {
-                        match listener.accept() {
-                            Ok((mut stream, _)) => {
-                                let connection_token = Token(token_id);
-                                println!("New connection on {}", listener.local_addr().unwrap());
-                                token_id += 1;
-
-                                poll.registry()
-                                    .register(&mut stream, connection_token, Interest::READABLE)
-                                    .expect("Failed to register new connection");
-
-                                connections.insert(
-                                    connection_token,
-                                    (stream, Arc::clone(&listener_config.config)),
-                                );
-                            }
-                            Err(ref e) if e.kind() == ErrorKind::WouldBlock => {
-                                // No more connections to accept
-                                break;
-                            }
-                            Err(e) => eprintln!("Error accepting connection: {}", e),
-                        }
-                    }
-                } else if let Some((stream, config)) = connections.get_mut(&token) {
-                    // Handle existing connection
-                    if let Err(e) = handle_client(stream, config) {
-                        if e.kind() != ErrorKind::WouldBlock {
-                            // Mark connection for closure
-                            println!("Marking connection for closure due to error: {}", e);
-                            to_close.push(token);
-                        }
-                    }
-                }
-            }
-            //  Close marked connections
-            for token in to_close {
-                if let Some((mut stream, _)) = connections.remove(&token) {
-                    poll.registry()
-                        .deregister(&mut stream)
-                        .expect("Failed to deregister stream");
-                }
-            }
-        }
     }
 }
