+++ src/server/start.rs
@@ -1,7 +1,11 @@
-use crate::server::{handle_client, Server};
+use crate::server::{handle_client, ListenerConfig, Server};
 use crate::server_config::server_config;
+use mio::net::TcpListener;
+use mio::{Events, Interest, Poll, Token};
+use std::collections::HashMap;
 use std::io::ErrorKind;
-use std::net::TcpListener;
+use std::net::SocketAddr;
+use std::sync::Arc;
 
 pub fn servers() -> Vec<Server<'static>> {
     let mut servers = Vec::new();
@@ -11,9 +15,8 @@ pub fn servers() -> Vec<Server<'static>> {
         for port in &config.ports {
             // Create a listener for each port
             let address = format!("{}:{}", config.host, port);
-            match TcpListener::bind(&address) {
+            match TcpListener::bind(address.parse::<SocketAddr>().unwrap()) {
                 Ok(listener) => {
-                    listener.set_nonblocking(true).unwrap();
                     listeners.push(listener);
                     println!("Server listening on {}", address);
                 }
@@ -26,24 +29,94 @@ pub fn servers() -> Vec<Server<'static>> {
     servers
 }
 
-pub fn start(mut servers: Vec<Server>) {
+pub fn start(servers: Vec<Server<'static>>) {
+    let mut poll = Poll::new().expect("Failed to create Poll instance");
+    let mut events = Events::with_capacity(128);
+    let mut connections = HashMap::new();
+    let mut token_id = 2; // Start token counting from 2
+    let mut all_listeners = Vec::new(); // Store all listeners
+    let mut listener_configs = Vec::new(); // Store ListenerConfig instances
+
+    for server in servers {
+        let config = Arc::new(server.config);
+
+        for listener in server.listeners {
+            let listener_index = all_listeners.len();
+            all_listeners.push(listener);
+            let token = Token(token_id);
+            token_id += 1;
+
+            poll.registry()
+                .register(
+                    &mut all_listeners[listener_index],
+                    token,
+                    Interest::READABLE,
+                )
+                .expect("Failed to register listener");
+
+            listener_configs.push(ListenerConfig {
+                listener_index,
+                config: Arc::clone(&config),
+            });
+        }
+    }
+
+    // Event loop
     loop {
-        for server in &mut servers {
-            for listener in &mut server.listeners {
-                match listener.accept() {
-                    Ok((mut stream, _addr)) => match handle_client(&mut stream, &server.config) {
-                        Ok(_) => continue,
-                        Err(e) => eprintln!("Error: {e}"),
-                    },
-                    Err(ref e) if e.kind() == ErrorKind::WouldBlock => {
-                        // No incoming connections, continue to the next listener
-                        continue;
+        poll.poll(&mut events, None).expect("Poll failed");
+        let mut to_close = Vec::new();
+
+        for event in events.iter() {
+            let token = event.token();
+
+            if let Some(listener_config) = listener_configs
+                .iter()
+                .find(|lc| token == Token(lc.listener_index + 2))
+            {
+                let listener = &mut all_listeners[listener_config.listener_index];
+
+                // Accept new connections in a loop
+                loop {
+                    match listener.accept() {
+                        Ok((mut stream, _)) => {
+                            let connection_token = Token(token_id);
+                            println!("New connection on {}", listener.local_addr().unwrap());
+                            token_id += 1;
+
+                            poll.registry()
+                                .register(&mut stream, connection_token, Interest::READABLE)
+                                .expect("Failed to register new connection");
+
+                            connections.insert(
+                                connection_token,
+                                (stream, Arc::clone(&listener_config.config)),
+                            );
+                        }
+                        Err(ref e) if e.kind() == ErrorKind::WouldBlock => {
+                            // No more connections to accept
+                            break;
+                        }
+                        Err(e) => eprintln!("Error accepting connection: {}", e),
                     }
-                    Err(e) => eprintln!("Error accepting connection: {}", e),
+                }
+            } else if let Some((stream, config)) = connections.get_mut(&token) {
+                // Handle existing connection
+                if let Err(e) = handle_client(stream, config) {
+                    if e.kind() != ErrorKind::WouldBlock {
+                        // Mark connection for closure
+                        println!("Marking connection for closure due to error: {}", e);
+                        to_close.push(token);
+                    }
                 }
             }
         }
-        // Sleep for a short duration to avoid busy waiting
-        std::thread::sleep(std::time::Duration::from_millis(1));
+        //  Close marked connections
+        for token in to_close {
+            if let Some((mut stream, _)) = connections.remove(&token) {
+                poll.registry()
+                    .deregister(&mut stream)
+                    .expect("Failed to deregister stream");
+            }
+        }
     }
 }
