+++ tests/common.rs
@@ -1,7 +1,5 @@
 #![allow(dead_code)]
-
 use localhost::server::{servers, start};
-
 pub fn setup() {
     start(servers())
 }
